refactor(deck): Page/Play/Gallery カラムから詳細ビューを切り出してウィンドウ化
カラムは「アイテム一覧」、ウィンドウは「アイテム詳細」という役割分担に統一。
ノート列系カラム (Ch/User/Clip/List/Antenna) との一貫性を取りつつ、
将来プロフィールタブから同じ詳細ウィンドウを起動できる土台を作る。

- 新規 useAiScriptRunner composable: Page/Play で重複していた AiScript
  セットアップ (parse → env → ui → interpreter → exec + Nd:* cleanup) を
  1 箇所に集約
- 新規 PageDetailContent / PlayDetailContent / GalleryDetailContent ウィンドウ
- DeckPageColumn / DeckPlayColumn / DeckGalleryColumn は一覧表示のみに縮退
  (合計 1500 行以上を削除)、アイテムクリックで対応ウィンドウを起動
